Add --only-my-changes flag to breeze static-checks command
When you have a workflow, when you rebase several of your commits,
on top of main, there is often a need to re-run pre-commits for
all the commits that have been changed by me in the branch (this
for example handles case when pre-commits changed in the main branch,
but the rebase did not result in conflict (this does not run pre-commit
automatically even if you have it installed in your repository).

It's also a good "sanity" check to run before you push your branch.