Fix broken subnet config test
The subnet config test will fail when run with the other tests in the
module (e.g. just running pytest on the whole test module). This is
because python only loads a module once and the ecs config module runs
its code and sets its values at the module level, so it's behaviour is
unchanged for the remainder of the python session. So the module must be
reloaded to test any changes.