Fix ecs attribute on ECS Executor
The attribute is expected to be a boto client, but was recently
changed to be a hook, which caused all api requests to fail. Updated the
attribute to be the Hook's conn (which is all we care about).

While I was there I moved all the intialization from start() to
__init__(). I don't see any reason we need to keep them separate