When clearing a task, clear the executor field on TI DB records
If a user tags a task to run with a particular executor, then triggers a
dag run, the task instance created persists the requested executor in
the DB record for that TI. However, if the user made a mistake in the
executor (typo, executor doesn't exist, changes their mind, etc) this
change allows them to reset that field by clearing the task.

The task instance will be updated with the latest executor tagged for
the task in the dag upon clearing.