Update ECS executor to support Task SDK
Follows the paradigm used in the k8s executor except the serialized
workload is passed along as string instead of a file, since there
is no (easy) way to plumb a file into the ECS container at run time.

In the future using a secret to pass along the workload could be
another possible approach, but this requires the use of another
cloud service (i.e. Secret Manager or Parameter Store).