DO NOT MERGE - AIP-86 - Fix callback serialization
see https://app.asana.com/1/8442528107068/project/1208344564272308/task/1210892497087168?focus=true

After some research, the Triggerer can not have access to the dagbag.  This was an intentional decision by the community.  If we want to do this we need to figure out how dagbundles work, and make this work with dag versions via dag bundles.  Jed says this is going to be a lot of effort and may not be worthwhile.

Slack chat in case it's still around:  https://apache-airflow.slack.com/archives/C07813CNKA8/p1753743831480199

Saving this commit in case it helps later if we decide to try to tackle that dagbundle update, but this should not be pushed as-is.