[v3-0-test] Fix remote logging connection availability in Task SDK supervisor (#54679)
Using connections stored in the Airflow Metadata for Remote logging was partially fixed
in 3.0.4, but the connection was only set/put-in-scope when the logger was created before
task startup. However for blob stores (i.e. S3 or WASB) that is a no-op, and we needed the
connection when we try to upload too.
(cherry picked from commit b340e8c101c1ed12fb097f9aa5b41bb9c2599275)

Co-authored-by: Rahul Vats <43964496+vatsrahul1001@users.noreply.github.com>