Fix connection management for EKS token generation
EKS token generation happens as a result of a bash script inserted into
the kube config and called by k8s code. This script cannot access
connections and variables since it does not communicate with the Task
SDK API. So instead create a closure of credentials and information
needed to create the token and embed it in the script.